+++ service/src/main/kotlin/fi/espoo/vekkuli/service/BoatSpaceService.kt
@@ -1,7 +1,9 @@
 package fi.espoo.vekkuli.service
 
 import fi.espoo.vekkuli.domain.*
+import fi.espoo.vekkuli.utils.mToCm
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 
 data class BoatSpaceFilter(
     val boatType: BoatType? = null,
@@ -16,10 +18,30 @@ interface BoatSpaceRepository {
     fun getUnreservedBoatSpaceOptions(params: BoatSpaceFilter): Pair<List<Harbor>, Int>
 }
 
+fun <T> getSingleOrEmptyList(item: T?): List<T> = if (item != null) listOf(item) else listOf()
+
 @Service
 class BoatSpaceService(
     private val boatSpaceRepo: BoatSpaceRepository
 ) {
-    fun getUnreservedBoatSpaceOptions(params: BoatSpaceFilter): Pair<List<Harbor>, Int> =
-        boatSpaceRepo.getUnreservedBoatSpaceOptions(params)
+    fun getUnreservedBoatSpaceOptions(
+        boatType: BoatType? = null,
+        width: BigDecimal? = null,
+        length: BigDecimal? = null,
+        amenities: List<BoatSpaceAmenity>? = null,
+        storageType: BoatSpaceAmenity? = null,
+        boatSpaceType: BoatSpaceType? = null,
+        harbor: List<String>? = null,
+    ): Pair<List<Harbor>, Int> {
+        val params =
+            BoatSpaceFilter(
+                boatType,
+                width?.mToCm(),
+                length?.mToCm(),
+                if (boatSpaceType != BoatSpaceType.Storage) amenities else getSingleOrEmptyList(storageType),
+                boatSpaceType,
+                if (boatSpaceType != BoatSpaceType.Storage) harbor?.map { s -> s.toInt() } else null
+            )
+        return boatSpaceRepo.getUnreservedBoatSpaceOptions(params)
+    }
 }
